extract concern at_school_period and add validation of ASPs range
