+++ app/models/mentor_at_school_period.rb
@@ -1,17 +1,18 @@
 class MentorAtSchoolPeriod < ApplicationRecord
   include Interval
   include DeclarativeUpdates
+  include AtSchoolPeriod
 
   # Associations
   belongs_to :school, inverse_of: :mentor_at_school_periods
   belongs_to :teacher, inverse_of: :mentor_at_school_periods
+
   has_many :mentorship_periods, inverse_of: :mentor, dependent: :destroy
   has_many :current_or_future_mentorship_periods,
            -> { current_or_future },
            class_name: "MentorshipPeriod"
   has_many :training_periods, inverse_of: :mentor_at_school_period, dependent: :destroy
   has_many :declarations, through: :training_periods
-  has_many :events
   has_many :lead_provider_metadata_for_mentees,
            class_name: "Metadata::TeacherLeadProvider",
            foreign_key: :ect_assigned_mentor_latest_school_period_id,
@@ -21,50 +22,13 @@ class MentorAtSchoolPeriod < ApplicationRecord
            -> { current_or_future.induction_not_completed.includes(:teacher) },
            through: :current_or_future_mentorship_periods,
            source: :mentee
-  has_one :current_or_next_training_period, -> { current_or_future.earliest_first }, class_name: "TrainingPeriod"
-  has_one :earliest_training_period, -> { earliest_first }, class_name: "TrainingPeriod"
-  has_one :latest_training_period, -> { latest_first }, class_name: "TrainingPeriod"
 
-  touch -> { teacher }, on_event: %i[create destroy update], when_changing: %i[email], timestamp_attribute: :api_updated_at
   touch -> { teacher }, on_event: %i[create destroy update], when_changing: %i[email], timestamp_attribute: :api_unfunded_mentor_updated_at, if: :latest_mentor_at_school_period?
 
-  refresh_metadata -> { school }, on_event: %i[create destroy update]
-  refresh_metadata -> { teacher }, on_event: %i[create destroy update]
-
   # Validations
-  validates :email,
-            notify_email: true,
-            allow_nil: true
-
-  validates :started_on,
-            presence: true
-
-  validates :school_id,
-            presence: true
-
-  validates :teacher_id,
-            presence: true
-
   validate :teacher_school_distinct_period
 
   # Scopes
-  scope :for_school, ->(school_id) { where(school_id:) }
-  scope :for_teacher, ->(teacher_id) { where(teacher_id:) }
-  scope :with_partnerships_for_contract_period, ->(year) {
-    joins(training_periods: {
-      active_lead_provider: :contract_period
-    }).where(contract_periods: { year: })
-  }
-  scope :with_expressions_of_interest_for_contract_period, ->(year) {
-    joins(training_periods: {
-      expression_of_interest: :contract_period
-    })
-    .where(contract_periods: { year: })
-  }
-  scope :with_expressions_of_interest_for_lead_provider_and_contract_period, ->(year, lead_provider_id) {
-    with_expressions_of_interest_for_contract_period(year)
-    .where(expression_of_interest: { lead_provider_id: })
-  }
 
   # Instance methods
   def siblings
@@ -72,17 +36,6 @@ class MentorAtSchoolPeriod < ApplicationRecord
 
     teacher.mentor_at_school_periods.for_school(school_id).excluding(self)
   end
-
-  def reported_leaving_by?(school)
-    reported_leaving_by_school_id.present? && reported_leaving_by_school_id == school&.id
-  end
-
-  def leaving_reported_for_school?(school)
-    leaving_today_or_in_future? && reported_leaving_by?(school)
-  end
-
-  delegate :provider_led_training_programme?, to: :current_or_next_training_period, allow_nil: true
-  delegate :school_led_training_programme?, to: :current_or_next_training_period, allow_nil: true
 
 private
